Remade folder structure inside folder' components'
