+++ client/src/components/WarningsComponents/LowSugarTodo.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { lowLevels, treatingLBS } from "../research";
+import { lowLevels, treatingLBS } from "../../research";
 import { makeStyles } from "@material-ui/core/styles";
 import Card from "@material-ui/core/Card";
 import CardContent from "@material-ui/core/CardContent";
