+++ client/src/components/WarningsComponents/NormalLevelCard.js
