+++ client/src/pages/ProgressPage.js
@@ -1,13 +1,16 @@
 import React from "react";
 import BloodSugarCard from "../components/BloodSugarCard";
-import ChartCard from "../components/ChartCard";
-import { LBSSymptomsCard, HBSSymptomsCard } from "../components/SymptomsCard";
+import ChartCard from "../components/ChartComponents/ChartCard";
+import {
+  LBSSymptomsCard,
+  HBSSymptomsCard,
+} from "../components/SymptomsComponents/SymptomsCard";
 import CardGrid from "../components/CardGrid";
-import LowLevelCard from "../components/LowLevelCard";
-import HighLevelCard from "../components/HighLevelCard";
-import NormalLevelCard from "../components/NormalLevelCard";
+import LowLevelCard from "../components/WarningsComponents/LowLevelCard";
+import HighLevelCard from "../components/WarningsComponents/HighLevelCard";
+import NormalLevelCard from "../components/WarningsComponents/NormalLevelCard";
 import ComplicationsCard from "../components/ComplicationsCard";
-import CareScheduleAccordion from "../components/CareSchedule";
+import CareScheduleAccordion from "../components/CareScheduleComponents/CareSchedule";
 
 function ProgressPage(props) {
   return (
